store: Move PoolStateTracker and handlers to own module
Also, rename it to just StateTracker
